+++ cube.h
@@ -11,7 +11,7 @@ public:
     virtual ~Cube();
     float* getVertices();
     float* getNormals();
-    unsigned int* getIndices();
+    unsigned short int* getIndices();
     unsigned char* getColours();
     float* getCoords();
     void draw();
